+++ Api/Infrastructure/ServiceCollectionExtensions.cs
@@ -101,7 +101,7 @@ using HappyTravel.Edo.Api.Services.SupplierResponses;
 using HappyTravel.SuppliersCatalog;
 using IdentityModel.Client;
 using Prometheus;
-using HappyTravel.Edo.Api.Services.Hotel;
+using HappyTravel.Edo.Api.Services.PropertyOwners;
 
 namespace HappyTravel.Edo.Api.Infrastructure
 {
@@ -762,7 +762,7 @@ namespace HappyTravel.Edo.Api.Infrastructure
             services.AddTransient<IRecordManager<AgencyProductivity>, AgenciesProductivityRecordManager>();
             services.AddTransient<IFixHtIdService, FixHtIdService>();
 
-            services.AddTransient<IHotelConfirmationService, HotelConfirmationService>();
+            services.AddTransient<IBookingConfirmationService, BookingConfirmationService>();
 
             //TODO: move to Consul when it will be ready
             services.AddCurrencyConversionFactory(new List<BufferPair>
