+++ Api/Models/Hotels/BookingConfirmation.cs
@@ -3,11 +3,11 @@ using System;
 
 namespace HappyTravel.Edo.Api.Models.Hotels
 {
-    public readonly struct HotelConfirmation
+    public readonly struct BookingConfirmation
     {
         public string ReferenceCode { get; init; }
         public string ConfirmationCode { get; init; }
-        public HotelConfirmationStatuses Status { get; init; }
+        public BookingConfirmationStatuses Status { get; init; }
         public string Initiator { get; init; }
         public DateTime CreatedAt { get; init; }
     }
