+++ Api/Services/PropertyOwners/IBookingConfirmationService.cs
@@ -1,0 +1,11 @@
+﻿using CSharpFunctionalExtensions;
+using HappyTravel.Edo.Api.Models.Hotels;
+using System.Threading.Tasks;
+
+namespace HappyTravel.Edo.Api.Services.PropertyOwners
+{
+    public interface IBookingConfirmationService
+    {
+        Task<Result> Update(BookingConfirmation hotelConfirmation);
+    }
+}
